requirements.txt
# SecLogAI Django v2.0 Requirements

# Django
Django>=5.2.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Data Processing & ML
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# OpenAI Chatbot
openai>=1.0.0

# Utils
python-dotenv>=1.0.0